Use CesiumIon for ion hosted examples
Added CesiumIon.create3DTileset helper function because updating the
examples showed it to be a useful shortcut. Also make `create3DTileset`
and `createImageryProvider` wait for the `readyPromise` because it removes
an extra step for the end user.

Finally, cleaned up some weird `throw` calls in example error handling
which just hid the error from the end-user.  Replaced them with logs.